Create libhdbppConfig.cmake to find the module upon installation. Moved HdbClient.h to src as it is a private interface. modify hdbclient.cpp accordingly
