+++ src/HdbClient.h
